+++ src/components/InstancesList/InstancesList.js
@@ -41,10 +41,11 @@ import {
   DefaultMCLRowFormatter,
 } from '@folio/stripes/components';
 import {
+  advancedSearchQueryBuilder,
   FACETS,
   facetsStore,
   queryIndexes,
-} from "@folio/stripes-inventory-components";
+} from '@folio/stripes-inventory-components';
 
 import { withSingleRecordImport } from '..';
 import FilterNavigation from '../FilterNavigation';
@@ -276,6 +277,7 @@ class InstancesList extends React.Component {
     }
 
     const searchParams = new URLSearchParams(location.search);
+
     const filters = searchParams.get('filters');
 
     const staffSuppressFalse = `${FACETS.STAFF_SUPPRESS}.false`;
@@ -1235,25 +1237,6 @@ class InstancesList extends React.Component {
     const formattedSearchableIndexes = searchableIndexes.map(this.formatSearchableIndex);
 
     const advancedSearchOptions = advancedSearchIndexes[segment].map(this.formatSearchableIndex);
-
-    const advancedSearchQueryBuilder = (rows) => {
-      const formatRowCondition = (row) => {
-        // use default row formatter, but wrap each search term with parentheses
-
-        const query = `${row.searchOption} ${row.match} ${row.query}`;
-        return query;
-      };
-
-      return rows.reduce((formattedQuery, row, index) => {
-        const rowCondition = formatRowCondition(row);
-
-        if (index === 0) {
-          return rowCondition;
-        }
-
-        return `${formattedQuery} ${row.bool} ${rowCondition}`;
-      }, '');
-    };
 
     const shortcuts = [
       {