+++ src/components/ViewSource/ViewSource.js
@@ -1,15 +1,29 @@
 import React, {
   useState,
   useEffect,
+  useMemo,
+  useCallback,
 } from 'react';
+import {
+  useLocation,
+  useHistory,
+} from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { FormattedMessage } from 'react-intl';
+import {
+  useIntl,
+  FormattedMessage,
+} from 'react-intl';
 
 import {
   Button,
   LoadingView,
+  HasCommand,
+  checkScope,
 } from '@folio/stripes/components';
-import { useStripes } from '@folio/stripes/core';
+import {
+  useCallout,
+  useStripes,
+} from '@folio/stripes/core';
 import {
   MarcView,
   PrintPopup,
@@ -18,7 +32,11 @@ import {
 
 import { useGoBack } from '../../common/hooks';
 
-import { isUserInConsortiumMode } from '../../utils';
+import {
+  isUserInConsortiumMode,
+  handleKeyCommand,
+  redirectToMarcEditPage,
+} from '../../utils';
 import MARC_TYPES from './marcTypes';
 
 import styles from './ViewSource.css';
@@ -32,7 +50,11 @@ const ViewSource = ({
   tenantId,
   marcType,
 }) => {
+  const intl = useIntl();
   const stripes = useStripes();
+  const location = useLocation();
+  const history = useHistory();
+  const callout = useCallout();
   const [isShownPrintPopup, setIsShownPrintPopup] = useState(false);
   const openPrintPopup = () => setIsShownPrintPopup(true);
   const closePrintPopup = () => setIsShownPrintPopup(false);
@@ -50,6 +72,31 @@ const ViewSource = ({
 
   const [marc, setMarc] = useState();
   const [isMarcLoading, setIsMarcLoading] = useState(true);
+
+  const redirectToMARCEdit = useCallback(() => {
+    const urlId = isHoldingsRecord ? `${instanceId}/${holdingsRecordId}` : instanceId;
+    const pathname = `/inventory/quick-marc/edit-${isHoldingsRecord ? 'holdings' : 'bib'}/${urlId}`;
+
+    redirectToMarcEditPage(pathname, instance, location, history);
+  }, [isHoldingsRecord]);
+
+  const shortcuts = useMemo(() => [
+    {
+      name: 'editMARC',
+      handler: handleKeyCommand(() => {
+        if ((marcType === MARC_TYPES.BIB && !stripes.hasPerm('ui-quick-marc.quick-marc-editor.all'))
+        || (marcType === MARC_TYPES.HOLDINGS && !stripes.hasPerm('ui-quick-marc.quick-marc-holdings-editor.all'))) {
+          callout.sendCallout({
+            type: 'error',
+            message: intl.formatMessage({ id: 'ui-inventory.shortcut.editMARC.noPermission' }),
+          });
+          return;
+        }
+
+        redirectToMARCEdit();
+      }),
+    },
+  ], [stripes, redirectToMARCEdit]);
 
   useEffect(() => {
     setIsMarcLoading(true);
@@ -94,32 +141,38 @@ const ViewSource = ({
   );
 
   return (
-    <div className={styles.viewSource}>
-      <MarcView
-        paneTitle={paneTitle}
-        marcTitle={marcTitle}
-        marc={marc}
-        onClose={goBack}
-        lastMenu={
-          isPrintAvailable &&
-            <Button
-              marginBottom0
-              buttonStyle="primary"
-              onClick={openPrintPopup}
-            >
-              <FormattedMessage id="ui-quick-marc.print" />
-            </Button>
-        }
-      />
-      {isPrintAvailable && isShownPrintPopup && (
-        <PrintPopup
-          marc={marc}
-          paneTitle={instance.title}
+    <HasCommand
+      commands={shortcuts}
+      isWithinScope={checkScope}
+      scope={document.body}
+    >
+      <div className={styles.viewSource}>
+        <MarcView
+          paneTitle={paneTitle}
           marcTitle={marcTitle}
-          onAfterPrint={closePrintPopup}
+          marc={marc}
+          onClose={goBack}
+          lastMenu={
+            isPrintAvailable &&
+              <Button
+                marginBottom0
+                buttonStyle="primary"
+                onClick={openPrintPopup}
+              >
+                <FormattedMessage id="ui-quick-marc.print" />
+              </Button>
+          }
         />
-      )}
-    </div>
+        {isPrintAvailable && isShownPrintPopup && (
+          <PrintPopup
+            marc={marc}
+            paneTitle={instance.title}
+            marcTitle={marcTitle}
+            onAfterPrint={closePrintPopup}
+          />
+        )}
+      </div>
+    </HasCommand>
   );
 };